GRU model, train on BPR loss

+++ model/loss.py
@@ -36,10 +36,11 @@ def neg_sample(packed_sess):
 def bpr_loss(y_input, y_target, n_samp=4):
     # getting tensors, batch size from packed sequences
     score, batch_sizes = y_input
+
     target, _ = y_target
     batch_size = batch_sizes[0]
 
-    loss_batch = torch.zeros((batch_size, 1), dtype=torch.float64)
+    loss_batch = torch.zeros((batch_size, 1), dtype=torch.float32)
     for i in range(batch_size):
         pos_score = torch.masked_select(score, target.byte())
         for j in range(n_samp):
@@ -48,7 +49,7 @@ def bpr_loss(y_input, y_target, n_samp=4):
             loss_sample = - torch.sum(F.logsigmoid(pos_score - neg_score)) / n_samp
             loss_batch[i] += loss_sample
             
-    return torch.sum(loss_batch) / batch_size.double()
+    return torch.sum(loss_batch) / batch_size.float()
 
     
 def top1_loss(y_input, y_target, n_samp=4):
